+++ weblate/machinery/views.py
@@ -423,7 +423,7 @@ def handle_machinery(request: AuthenticatedHttpRequest, service, unit, search=No
         except MachineTranslationError as exc:
             response["responseDetails"] = str(exc)
         except Exception as error:
-            report_error(project=component.project)
+            report_error("Machinery failed", project=component.project)
             response["responseDetails"] = f"{error.__class__.__name__}: {error}"
 
     if response["responseStatus"] != 200: